Moved upsat related files to upsate folder
